+++ src/main/java/net/wizards/item/WizardArmor.java
@@ -1,51 +1,11 @@
 package net.wizards.item;
 
-import mod.azure.azurelibarmor.common.api.client.renderer.GeoArmorRenderer;
-import mod.azure.azurelibarmor.common.api.common.animatable.GeoItem;
-import mod.azure.azurelibarmor.common.internal.client.RenderProvider;
-import mod.azure.azurelibarmor.common.internal.common.util.AzureLibUtil;
-import mod.azure.azurelibarmor.core.animatable.instance.AnimatableInstanceCache;
-import mod.azure.azurelibarmor.core.animation.AnimatableManager;
-import net.minecraft.client.render.entity.model.BipedEntityModel;
-import net.minecraft.entity.EquipmentSlot;
-import net.minecraft.entity.LivingEntity;
 import net.minecraft.item.ArmorMaterial;
-import net.minecraft.item.ItemStack;
 import net.minecraft.registry.entry.RegistryEntry;
 import net.spell_engine.api.item.armor.Armor;
-import net.wizards.client.armor.WizardArmorRenderer;
 
-import java.util.function.Consumer;
-
-public class WizardArmor extends Armor.CustomItem implements GeoItem {
+public class WizardArmor extends Armor.CustomItem {
     public WizardArmor(RegistryEntry<ArmorMaterial> material, Type slot, Settings settings) {
         super(material, slot, settings);
-    }
-
-    // MARK: GeoItem
-
-    @Override
-    public void createRenderer(Consumer<RenderProvider> consumer) {
-        consumer.accept(new RenderProvider() {
-            private GeoArmorRenderer<?> renderer;
-            @Override
-            public BipedEntityModel<LivingEntity> getHumanoidArmorModel(LivingEntity livingEntity, ItemStack itemStack, EquipmentSlot equipmentSlot, BipedEntityModel<LivingEntity> original) {
-                if (this.renderer == null) {
-                    this.renderer = new WizardArmorRenderer();
-                }
-                this.renderer.prepForRender(livingEntity, itemStack, equipmentSlot, original);
-                return this.renderer;
-            }
-        });
-    }
-
-    @Override
-    public void registerControllers(AnimatableManager.ControllerRegistrar controllers) { }
-
-    private final AnimatableInstanceCache cache = AzureLibUtil.createInstanceCache(this);
-
-    @Override
-    public AnimatableInstanceCache getAnimatableInstanceCache() {
-        return cache;
     }
 }